Changed the MonodomainDg0Assembler tests to use 1mm rather than 1cm meshes, with a 0.1mm space step.
Moved TestMonodomainSlab to the nightly test pack so it gets run, and
included a test on a 1mm x 1mm x 1mm mesh.

Note that tests in TestMonodomainDg0Assembler have large tolerances, due to
measuring transmembrane potential during the upstroke, and slight
irregularities in triangle orientation leading to substantial variability.
This should be reduced if a finer mesh is used (according to JonW).