Some modifications to build scripts to build on comlab client PCs.
Refactoring to improve performance loading meshes with internal faces.
Culling internal faces is no longer done by the *MeshReader object; rather the
ConformingTetrahedralMesh method to read from a MeshReader works out which faces are on the
boundary.

To aid this, nodes now contain a set of pointers to elements that contain them as a vertex.