Work related to ticket #135
The coverage build can now recognise some spurious failures, i.e. lines that are reported as not executed
but almost certainly are.  The 2 classes it checks for are
 1) lines with just a closing brace
 2) lines with just a return statement

The test summary scripts can now display 3 summary colours: red, orange and green.  If a file only has
spurious failures under coverage testing, it is shown as orange.