Fixed Lee's memory leak (deleted object)
Made big improvement to 1D/2D mesh in 2D/3D space tests.  Formally our meshes were closed (circle in 2D or the surface of a watertight solid 
in 3D).  Now we have some open meshes.

In 1Din2D the nodes are culled to find the boundary (either 0 or 2 boundary elements).

The new rule is that a 2D mesh in 3D must give .node .face '''and .edge''' files.  The edges are
automatically culled to find the boundary.