Reverted TestBidomainHeart.hpp.
changeset:1137 had an accidental commit of Test[Mono/Bi]domainHeart.hpp, whilst they were still work 
progress (trying to get them to work with the full heart mesh). TestMonodomainHeart 
has been fixed but reverting TestBidomainHeart for the time being. 

Reverted to revision 1092 except redid changes in changeset:1164, and changed ints to unsigned in 
FinaliseCellCreation(). FinaliseCellCreation() was not actually being called in revison 1092 because 
of this issue.