ticket #163
We believe to have got the algorithm for refining an element right, but the implementation seems to be a bit 
tricky. One should have a look at RefineElement in ConformingTetrahedralMesh.cpp. The problem seems to be when 
we try to push the new element at the end of the vector (step #3). If we comment out the code that adds a new 
node then it's kind of fine... Weird!