#767 - Using jacobian determinant cached in mesh class.
`RefreshJacobianCachedData` can now cope with a mesh that changes size, and also with deleted elements.

Note: we had to change `TissueSimulationWithNutrients` in order to get this to work.  That class
uses the cancer mesh as the mesh for a PDE solve in its `PostSolve` method.  Unfortunately, at the
point `PostSolve` is called the cancer mesh can be in an inconsistent state, as the node positions
have been updated without a remesh happening (a remesh is done on the next timestep, after death &
birth).  So we added an extra remesh in `TissueSimulationWithNutrients::SolveNutrientPde`.  There
might be a better way of doing it...