#1156 - change how executable tries schemas.  It now tries the hardcoded location first; if this schema isn't found, then the error message is swallowed and the schema location given in the XML file is tried.
This was done to fix the acceptance tests, since their parameters files don't specify a schema location at all.  When this happens, xerces
just gives validation errors assuming that the input should be empty, it doesn't complain about a missing schema!  So I had no nice way to
determine that we should try the hardcoded schema location in this instance.