#1162 - add support for derived quantities to pycml, using the pycml:derived-quantity annotation.
As part of this, extracted a new file `VectorHelperFunctions.hpp` from `AbstractParameterisedSystem.cpp`.

Also altered the names added to the ode system information, so if the variable has a cmeta:id this will be used in preference to its name.

Also re-generated backward Euler LR91 to test with parameter & derived quantities,
and fixed a bug in this related to work on #1354.