Tiket 1528: Editing the Installation guide.
In this version we change line where we build the optimized version of PETSC3. Parmetis library was not built in the optimized version of Petsc3.This is fixed now. We also correct the PYTHONPATH that should be set in the bachrc file, before starting compiling chaste.This installation guide was successfully tested on the clpc98 machine.