#1757 - improve performance of calculating APDs for many nodes by reading multiple columns at once from file.  * The low-level support is in `Hdf5DataReader::GetVariableOverTimeOverMultipleNodes`, which currently only works for complete data files (i.e. not ones which miss out some nodes).  * The APD support is in `PropagationPropertiesCalculator::CalculateAllActionPotentialDurationsForNodeRange`, which is now used by `PostProcessingWriter::WriteApdMapFile`.
Similar optimisations could be applied for the other methods in `PostProcessingWriter` if desired, in which case it might be best to abstract much of the logic in `CalculateAllActionPotentialDurationsForNodeRange` into a helper method, rather than replicating it for each case.

There's probably also scope for not moving data around in memory so much once it has been read from file!