Avoid an occasional error when building in parallel, whether with -j or a parallel test using dynamically loaded cell models.
There was a race condition in the !PyCml emitter which could try to figure out the
targets being generated for a dynamically loaded cell model being created by another
process, but find the other process had removed the source from under it and hence
fall over.

Done for #1795, but general fix.