#2394 - add the theoretical ability to debug python tests.
Unfortunately it's not very helpful since normally test output is stored in a Python buffer and written to file as well as screen, and so interactive debugging is impossible.
But you can set up the correct PYTHONPATH and CHASTE_DEBUG=1 in your terminal, then run python/infra/TestPythonCode.py path_to_test.py.