Fix nightly tests broken by r23351 (#2565).
The issue was that `FakePetscSetup` registered a citation for Chaste during the common PETSc initialisation,
then changed to be a single process.  However, the citation registration cached the fact that it was run in parallel,
so later checks of parallel status gave the wrong information.  `FakePetscSetup` now resets the cache once it
enters serial mode.